move validator package to top level
